pyproject.toml
[tool.poetry]
name = "xcrypt-python"
version = "0.1.0"
description = "Xcrypt parallel job scheduling tool"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "xcrypt_python"}]

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0"
pyyaml = "^6.0"
pytest = "^7.0"
rich = "^13.0"
mypy = "^1.14.1"
black = "^25.1.0"
flake8 = "^7.1.1"
isort = "^6.0.0"
openai = "^1.63.0"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
xcrypt-python = "xcrypt_python.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
